LAWN-04 : Enable SpiCommand protocol
While at it:
* Added getter to ps4 controller
* Extended README
* Dropped the example code for spi

Open:
* Tx buffer is set to 0 for master
* Slave does not properly store bytes
* Slave does not interpret bytes yet

+++ lawnmover_utils/serial_logger.cpp
@@ -90,7 +90,7 @@ void SerialLogger::log(const char *format, va_list argptr) {
                 } else if ( *c == 'c' ) {
                     Serial.print(va_arg( argptr, int ));
                 } else if ( *c == 'f' ) {
-                    Serial.print(va_arg( argptr, float ), 6);
+                    Serial.print(va_arg( argptr, double ), 6);
                 }
             } else {
                 Serial.print(*c);
